Add `vcr` testing framework and example unittests for `get_public_timeline`
Please make sure that you have the envvar `RTOOT_DEFAULT_TOKEN` which
contains your default token in plain text, IF YOU WANT TO RECORD NEW
VCR CASSETTES.

See `tests/testthat/setup-rtoot.R`.

If you don't set that up first, your token will be exposed in the
recorded cassette. If there are new cassettes in `fixtures` in
future commits, we need to check if any token is exposed.

However, if you just want to run those tests, you don't need to do
that. (But it doesn't harm to do that.)

I will also add an entry in our wiki about creating tests, if this is merged.

https://github.com/schochastics/rtoot/wiki