Make testing robust without the default token
If you just want to run those tests, you don't need to have the token.
A fake token is created at the beginning of the testing session. You
don't even need to have that envvar.

Have checked this without the default token at the default location;
as well as without the envvar.

However, if you want to create new tests, you need also to take this
fake token approach. We can't create tests with `token = NULL` because
different token objects point to different default instances. For
testing, we need to fix the instance.

It doesn't need to be "social.tchncs.de". But just need to fix to the
one you create the tests.